refactor(dfloat): split dfloat into core / manipulators / iostream (#1334)
Second type of Phase 2 group 5a (#1444), and it needed a substrate split first.

    core.hpp        58,787 lines / 295 headers / 0 I/O-family
    dfloat.hpp      76,512 lines / 377 headers / 5 I/O-family  (was 77,924 / 390 / 5)

Zero I/O-family headers, the binary criterion, is met. The under-45,000-lines
criterion is NOT, and reporting the number rather than moving the bar: the two
shared substrates a dfloat core cannot do without -- native/ieee754_core.hpp at
43,486 and internal/blockbinary/to_decimal.hpp at 33,308 -- already total 44,846
together, before dfloat contributes a line. That ceiling belongs to the
substrate, not to this type. For scale, the merged cores sit at cfloat 78,330,
fixpnt 78,123, posit 77,430, lns 74,762.

Substrate: internal/blockbinary/to_decimal.hpp (new)

to_decimal() builds its result by prepending characters to a std::string and
never opens a stream, but it shared manipulators.hpp with type_tag, to_binary and
to_hex, which do -- 50,808 lines and four of the five I/O-family headers. dfloat's
str() renders its significand with to_decimal(), and str() is core by the layer
contract, so the core could not reach zero while that function was trapped in a
streaming header. Split out: 33,308 lines, 0 I/O. manipulators.hpp includes it, so
every existing caller is unaffected. This is the rational precedent (#1425) taken
one step further -- rational could move its only consumer out of the core, dfloat
cannot.

What moved in dfloat:

- to_binary() and to_native() format through a stringstream -> manipulators.hpp.
- operator<< and operator>> -> iostream.hpp, which includes only core.hpp:
  operator<< formats through the core member str() and operator>> hands its token
  to parse(), so unlike bfloat16 there is no dependency on manipulators at all.
- parse() STAYS in the core. Unlike cfloat's and bfloat16's, dfloat's parse() is a
  pure character-grammar validator ending in value.assign(number) -- no
  istringstream anywhere -- so it is arithmetic surface, not text.
- native/ieee754.hpp -> native/ieee754_core.hpp.
- DFLOAT_ENABLE_LITERALS, DFLOAT_THROW_ARITHMETIC_EXCEPTION, DFLOAT_EXCEPT and
  DFLOAT_NATIVE_SQRT now default in dfloat_impl.hpp beside the code they govern,
  so the core-only path gets the same defaults the umbrella supplied (#1436).

dfloat_impl.hpp had handed blockbinary's type_tag/to_binary/to_hex to every dfloat
translation unit through its manipulators include. The core now takes only
to_decimal.hpp, so dfloat/manipulators.hpp picks that include up and the
umbrella's surface is unchanged.

manipulators.hpp and attributes.hpp had no includes at all -- they reached
std::stringstream, std::setw, Color, type_tag and operator<< entirely through the
umbrella's include order. Both now name what they use (#1389).

Verification:

- Layer gate: a TU including core.hpp alone, and one including
  core.hpp + manipulators.hpp, both compile, link and run under -Wall -Wpedantic
  on gcc 13.3 and clang.
- Self-containment: core, manipulators, iostream, attributes, the umbrella, and
  both blockbinary headers each compile standalone, 0 errors, both compilers.
  dfloat_impl.hpp and the two block test-suite headers still fail standalone --
  identically on main (dfloat_impl 254 errors there, 208 here), the known
  guarded-but-not-self-contained class, not a regression.
- 25 targets built and run green on both compilers: the 14 dfloat tests, plus the
  blockbinary and to_decimal consumers the substrate split reaches (bb_api,
  bb_conversion, bb_logic, bb_bit_patterns, bb_decimal_converter, bt_decimal,
  conversions_decimal_boundary, conversions_to_decimal,
  play_float_to_decimal_string), plus multifile and ucalc.
- Every header in the tree that includes blockbinary/manipulators.hpp or
  to_decimal.hpp syntax-checked on both compilers.
- Differential against main: all five standard aliases (decimal32/64/128 and the
  two DPD variants) over 64 encodings each through to_binary, to_native,
  color_print, components, str, operator<< under five stream states, plus
  parse/assign/operator>>/string-ctor over 29 decimal texts including every
  failure path, plus blockbinary to_decimal over 81 signed values and a 128-bit
  shifting sweep. 884,204 bytes of stdout and all 20 stderr diagnostics
  byte-identical, on gcc and clang.
