+++ include/sw/universal/internal/blockbinary/manipulators.hpp
@@ -6,16 +6,21 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 //
-// The <iomanip> half of blockbinary's text layer (#1334): type_tag, to_binary,
-// to_hex, to_decimal -- everything that returns a std::string. blockbinary.hpp
-// holds the arithmetic and needs none of it, which is what keeps <sstream> and
-// <iomanip> out of every number system that builds on blockbinary.
+// The <iomanip> half of blockbinary's text layer (#1334): type_tag, to_binary and
+// to_hex -- the string producers that format through a std::stringstream.
+// blockbinary.hpp holds the arithmetic and needs none of it, which is what keeps
+// <sstream> and <iomanip> out of every number system that builds on blockbinary.
+// to_decimal() concatenates rather than streams and now lives in to_decimal.hpp,
+// included below so this header's surface is unchanged.
 #include <cstdint>
 #include <iomanip>
 #include <sstream>
 #include <string>
 #include <typeinfo>
 #include <universal/internal/blockbinary/blockbinary.hpp>
+#include <universal/internal/blockbinary/to_decimal.hpp>   // to_decimal moved out (#1334): it
+                                                           // concatenates a string and needs no
+                                                           // stream, so a core can include it alone
 
 namespace sw { namespace universal {
 // Generate a type tag for blockbinary
@@ -62,48 +67,6 @@ std::string to_hex(const blockbinary<nbits, BlockType, NumberType>& number, bool
 		if (nibbleMarker && n > 0 && ((n * 4ll) % bitsInBlock) == 0) ss << '\'';
 	}
 	return ss.str();
-}
-
-// decimal string conversion
-template<unsigned nbits, typename BlockType, BinaryNumberType NumberType>
-std::string to_decimal(const blockbinary<nbits, BlockType, NumberType>& number) {
-	if (number.iszero()) return "0";
-
-	std::string result;
-	blockbinary<nbits, BlockType, NumberType> dividend(number);
-	bool isNegative = false;
-
-	// Handle negative numbers for signed types
-	if constexpr (NumberType == BinaryNumberType::Signed) {
-		if (dividend.isneg()) {
-			isNegative = true;
-			dividend.twosComplement(); // Convert to positive
-		}
-	}
-
-	// Repeatedly divide by 10 and collect remainders
-	blockbinary<nbits, BlockType, NumberType> ten(10);
-	while (!dividend.iszero()) {
-		if constexpr (nbits <= 64) {
-			// For smaller sizes, use native division to avoid complexity
-			uint64_t temp = dividend.to_ull();
-			uint64_t remainder = temp % 10;
-			result = char('0' + remainder) + result;
-			dividend = temp / 10;
-		} else {
-			// For larger sizes, use blockbinary division operators
-			blockbinary<nbits, BlockType, NumberType> remainder = dividend % ten;
-			uint64_t digit = remainder.to_ull();
-			result = char('0' + digit) + result;
-			dividend /= ten;
-		}
-	}
-
-	if (isNegative) {
-		result = "-" + result;
-	}
-
-	return result;
 }
 
 }} // namespace sw::universal
